Separate EnumComboBoxEditor class from ItemsListView
